requirements.txt
Flask>=2.0
Pillow>=9.0
numpy>=1.21
scikit-learn>=1.0
webcolors>=1.11
pytesseract>=0.3.10